requirements.txt
praw
python-dotenv
textblob
